完整源码/requirements.txt
flask
flask-sqlalchemy
flask-debugtoolbar
flask-sqlacodegen
mysqlclient
flask_script
requests
uwsgi